Refactor FactoryTest as real test
